requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
mysql-connector-python==8.0.32
protobuf==3.20.3
PyMySQL==1.0.2
requests==2.28.2
schedule==1.1.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.15.0
